requirements.txt
streamlit>=1.36.0
torch>=2.1.0
torchvision>=0.16.0
pillow
numpy
matplotlib
scikit-learn